Solution to Toll Plaza Management System

+++ Toll Plaza Management System@@67877d61dc107486ad5c2750/code.c
@@ -6,7 +6,7 @@ int main(){
     char number[n][10];
     char type[n][10];
     int tax[n];
-    int car_tax=0,truck_tax=0,bike_tax=0;
+    float car_tax=0,truck_tax=0,bike_tax=0;
     for(int i=1;i<=n;i++){
         scanf("%s %s %d",number[i],type[i],&tax[i]);
         if(type[i]=="Car"){
